requirements.txt
bidict==0.23.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
eventlet==0.33.3
flask==2.0.3
flask-cors==3.0.10
flask-socketio==4.3.2
greenlet==3.2.3
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-engineio==3.13.2
python-socketio==4.6.1
simple-websocket==1.1.0
six==1.17.0
Werkzeug==2.0.3
wsproto==1.2.0
pycryptodome
dtls